fix: passing static
