+++ discord-impl/src/main/java/net/starype/quiz/discordimpl/command/CommandContext.java
@@ -3,27 +3,20 @@ package net.starype.quiz.discordimpl.command;
 import net.dv8tion.jda.api.entities.Member;
 import net.dv8tion.jda.api.entities.Message;
 import net.dv8tion.jda.api.entities.TextChannel;
-import net.starype.quiz.discordimpl.util.DiscordContext;
 import net.starype.quiz.discordimpl.game.GameList;
 import net.starype.quiz.discordimpl.game.LobbyList;
 
 public class CommandContext {
 
-    private final DiscordContext discordContext;
     private MessageContext messageContext;
     private GameList gameList;
     private LobbyList lobbyList;
 
 
-    public CommandContext(MessageContext messageContext, GameList gameList, LobbyList lobbyList, DiscordContext discordContext) {
+    public CommandContext(MessageContext messageContext, GameList gameList, LobbyList lobbyList) {
         this.messageContext = messageContext;
-        this.discordContext = discordContext;
         this.gameList = gameList;
         this.lobbyList = lobbyList;
-    }
-
-    public DiscordContext getDiscordContext() {
-        return discordContext;
     }
 
     public TextChannel getChannel() {
