+++ discord-impl/src/main/java/net/starype/quiz/discordimpl/command/GenerateDatabaseCommand.java
@@ -5,7 +5,6 @@ import net.dv8tion.jda.api.entities.Message;
 import net.dv8tion.jda.api.entities.Message.Attachment;
 import net.dv8tion.jda.api.entities.TextChannel;
 import net.starype.quiz.api.database.*;
-import net.starype.quiz.discordimpl.util.DiscordContext;
 import net.starype.quiz.discordimpl.util.InputUtils;
 
 import java.io.ByteArrayInputStream;
@@ -36,7 +35,7 @@ public class GenerateDatabaseCommand implements QuizCommand {
                 ? args[1]
                 : "database";
 
-        Optional<InputStream> database = generateFile(files.iterator().next().getUrl(), channel, context.getDiscordContext());
+        Optional<InputStream> database = generateFile(files.iterator().next().getUrl(), channel);
         if(database.isEmpty()) {
             return;
         }
@@ -46,8 +45,8 @@ public class GenerateDatabaseCommand implements QuizCommand {
                 .queue();
     }
 
-    private static Optional<InputStream> generateFile(String urlName, TextChannel channel, DiscordContext discordContext) {
-        Collection<? extends EntryUpdater> updaters = InputUtils.loadEntryUpdaters(urlName, channel, discordContext.downloadingLimiter());
+    private static Optional<InputStream> generateFile(String urlName, TextChannel channel) {
+        Collection<? extends EntryUpdater> updaters = InputUtils.loadEntryUpdaters(urlName, channel);
         AtomicReference<ByteBuffer> output = new AtomicReference<>();
         SerializedIO serializedIO = new ByteSerializedIO(new byte[0], output);
         TrackedDatabase db = new QuestionDatabase(updaters, serializedIO, false);
